# [React] Enable Ahead-of-Time Compilation