@mcollina 2024-08-28
Unfortunately the point of this work is to analyze the impact of SSR. So it would make this essentially unfair.