@slorber 2024-08-28
Not exactly the same but I was wondering if the React compiler isn't also suited to optimize React SSR code by default? 

For example in this base app, most of the JSX is static can be shared between requests. Optimizing this manually is possible although not super natural, but I think it would be fair to plug a compiler that optimizes this automatically?

cc @gsathya is the React Compiler only for client-side optimizations? 

By that I mean something like this but more powerful: 

https://babeljs.io/docs/babel-plugin-transform-react-constant-elements.html


The following "optimized app" would increase perf by ~10-15%, but it's not realistic for human beings to optimize real production apps this way IMHO:

```jsx
const StaticAppElement = (function createStaticJSX() {
  const wrapperWidth = 960
  const wrapperHeight = 720
  const cellSize = 10
  const centerX = wrapperWidth / 2
  const centerY = wrapperHeight / 2

  let idCounter = 0
  let angle = 0
  let radius = 0

  const tiles = []
  const step = cellSize

  let x
  let y
  while (radius < Math.min(wrapperWidth, wrapperHeight) / 2) {
    x = centerX + Math.cos(angle) * radius
    y = centerY + Math.sin(angle) * radius

    if (x >= 0 && x <= wrapperWidth - cellSize && y >= 0 && y <= wrapperHeight - cellSize) {
      tiles.push({ x, y, id: idCounter++ })
    }

    angle += 0.2
    radius += step * 0.015
  }

  return (
      <div id="wrapper">
        {tiles.map((tile) => (
            <div
                key={tile.id}
                className="tile"
                style={{
                  left: tile.x,
                  top: tile.y,
                }} />
        ))}
      </div>
  )
})()

export function createApp () {
  return StaticAppElement
}
```


---

Side note: I also tried to increase perf by using `renderToStaticMarkup()` instead of `renderToString` but the impact is very small: https://github.com/platformatic/ssr-performance-showdown/pull/13