@gaearon 2024-08-28
I didn’t mean this PR too seriously, but I do want to point out that if Fastify positions itself as a minimal leaner alternative to frameworks, and as performance-conscious (to the point of doing benchmarks that recommend against using React), it would be good to point out where it’s behind state of the art performance optimizations that React frameworks can do. Pre-rendering static content at the build time is one of those optimizations, but I couldn’t find how to do that on the Fastify website. When looking at performance of different tools, I think it’s worthwhile to look at the big picture (how do rendering strategies affect performance, how do different frameworks let you shift work ahead of time, how they compare in terms of blocking on data vs streaming rendering while IO is loading) rather than only comparing the speed of emitting a thousand divs. This, IMO, would help readers develop a more nuanced and informed opinions.