@mcollina 2024-08-30
The point of the benchmark is not to recommend using React. In fact, we mostly use React at Platformatic.

Using a framework (or not) is often not dependent on the "rendering a thousand divs", but  sometimes is. That's what this benchmark is for.